+++ packages/parallax/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @smoovy/parallax
 
+## 1.0.5
+
+### Patch Changes
+
+- feat(parallax): added locks
+
+## 1.0.4
+
+### Patch Changes
+
+- feat(parallax): added target initial update
+
 ## 1.0.3
 
 ### Patch Changes